+++ src/Databases/DatabaseParentType.php
@@ -6,4 +6,5 @@ enum DatabaseParentType: string
 {
     case Page = "page_id";
     case Workspace = "workspace";
+    case Block = "block_id";
 }